chore: Organize scripts dir structure

+++ IIoTDashboard/scripts/deploy.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
 
+set -e
+
 APP_VERSION="1.0.0"
 APP_NAME="IIoTDashboard"
 APP_FOLDER="$APP_NAME-deploy"
-BUILD_PATH=build/WebAssembly_Qt_6_8_3_single_threaded-Debug
+BUILD_PATH=build-wasm
 
-rm -rf $APP_FOLDER
-rm *.deb
+bash scripts/build.sh -c
+
+# Remove previous deploy folder
+if [ -d $APP_FOLDER ]; then
+    rm -rf $APP_FOLDER
+fi
+
+# Remove previous .deb
+if [ -f $APP_NAME-$APP_VERSION.deb ]; then
+    rm $APP_NAME-$APP_VERSION.deb
+fi
 
 # Create the necessary paths to the application deploy
 mkdir $APP_FOLDER
@@ -14,12 +25,12 @@ mkdir -p $APP_FOLDER/DEBIAN
 mkdir -p $APP_FOLDER/etc/nginx/sites-available
 mkdir -p $APP_FOLDER/var/www/iiotdashboard
 
-
-cp control $APP_FOLDER/DEBIAN
-cp postinst $APP_FOLDER/DEBIAN
+# Copy files to deploy folder
+cp debian/control $APP_FOLDER/DEBIAN
+cp debian/postinst $APP_FOLDER/DEBIAN
 cp nginx/iiotdashboard $APP_FOLDER/etc/nginx/sites-available
-cp index.html $APP_FOLDER/var/www/iiotdashboard
-cp favicon.ico $APP_FOLDER/var/www/iiotdashboard
+cp www/index.html $APP_FOLDER/var/www/iiotdashboard
+cp www/favicon.ico $APP_FOLDER/var/www/iiotdashboard
 cp $BUILD_PATH/$APP_NAME.js $APP_FOLDER/var/www/iiotdashboard
 cp $BUILD_PATH/$APP_NAME.wasm $APP_FOLDER/var/www/iiotdashboard
 cp $BUILD_PATH/qtloader.js $APP_FOLDER/var/www/iiotdashboard
